+++ .vscode/settings.json
@@ -6,10 +6,13 @@
     "python.testing.unittestArgs": [
         "-v",
         "-s",
-        "./tests",
+        "./src",
         "-p",
-        "test_*.py"
+        "*test*.py"
     ],
     "python.testing.pytestEnabled": false,
-    "python.testing.unittestEnabled": true
+    "python.testing.unittestEnabled": true,
+    "python.testing.pytestArgs": [
+        "src"
+    ]
 }
