+++ src/rtichoke/__init__.py
@@ -5,23 +5,6 @@ from importlib.metadata import version
 
 # __version__ = version("rtichoke")
 
-# from rtichoke.discrimination.gains import create_gains_curve
-# from rtichoke.discrimination.gains import plot_gains_curve
-
-# from rtichoke.calibration.calibration import create_calibration_curve
-
-# from rtichoke.utility.decision import create_decision_curve
-# from rtichoke.utility.decision import plot_decision_curve
-
-# from rtichoke.performance_data.performance_data import prepare_performance_data
-# from rtichoke.performance_data.calculate_performance_data import PerformanceData
-
-# from rtichoke.summary_report.summary_report import create_summary_report
-
-# from .rtichoke import Rtichoke
-# from .performance_data._performance_data import *
-# from .helpers.validations import *
-
 from .helpers.helper_functions import *
 
 
@@ -32,7 +15,8 @@ class Rtichoke:
         prepare_performance_table,
     )
     from .helpers.validations import validate_inputs, validate_plot_inputs, check_by
-    from .plot._plot import plot, select_data_table
+    from .helpers.helper_functions import tprint, select_data_table
+    from .plot.plotting import plot
 
     def __init__(self, probs=None, reals=None, by=0.01):
         super().__init__()
@@ -46,14 +30,12 @@ class Rtichoke:
 
         self.check_by()
 
-        print_with_time(
-            "Calculating performance table stratified by probability threshold"
-        )
+        tprint("Calculating performance table stratified by probability threshold")
         self.performance_table_pt = self.prepare_performance_data(
             stratified_by="probability_threshold"
         )
 
-        print_with_time("Calculating performance table stratified by ppcr")
+        tprint("Calculating performance table stratified by ppcr")
         self.performance_table_ppcr = self.prepare_performance_data(
             stratified_by="ppcr"
         )
