+++ src/rtichoke/helpers/helper_functions.py
@@ -1,5 +1,20 @@
 from datetime import datetime
 
-def print_with_time(msg):
-    now = datetime.strftime(datetime.now(), '%Y-%m-%d %H:%M:%S')
-    print(now + ' - ' + msg)
+
+def tprint(string):
+    now = datetime.strftime(datetime.now(), "%Y-%m-%d %H:%M:%S")
+    print(now + " - " + string)
+
+
+def select_data_table(self, x, y, stratification="probability_threshold"):
+    df = (
+        self.performance_table_pt
+        if stratification == "probability_threshold"
+        else self.performance_table_ppcr
+    )
+    cols = list(
+        set(
+            ["Population", "predicted_positives", "probability_threshold", "ppcr", x, y]
+        )
+    )
+    return df[cols]
