+++ src/rtichoke/plot/bokeh/__init__.py
@@ -1,0 +1,3 @@
+"""
+Subpackage for bokeh plotting
+"""
