--- tests/test_rtichoke.py
@@ -1,10 +1,0 @@
-"""
-A module for tests
-"""
-
-# from rtichoke import rtichoke
-
-def test_addition():
-    expected = 4
-    actual = 2 + 2
-    assert actual == expected
